Replace int with int64_t

+++ torch/lib/THCUNN/generic/SpatialGridSamplerBilinear.cu
@@ -12,12 +12,12 @@ static inline void THNN_(SpatialGridSamplerBilinear_shapeCheck)(
   THCUNN_argCheck(state, THCTensor_(nDimension)(state, grid) == 4, 2, grid,
       "4D grid tensor expected but got: %s");
 
-  int nbatch   = THCTensor_(size)(state, input, 0);
-  int channels = THCTensor_(size)(state, input, 1);
-  int iheight   = THCTensor_(size)(state, input, 2);
-  int iwidth    = THCTensor_(size)(state, input, 3);
-  int oheight   = THCTensor_(size)(state, grid, 1);
-  int owidth    = THCTensor_(size)(state, grid, 2);
+  int64_t nbatch   = THCTensor_(size)(state, input, 0);
+  int64_t channels = THCTensor_(size)(state, input, 1);
+  int64_t iheight   = THCTensor_(size)(state, input, 2);
+  int64_t iwidth    = THCTensor_(size)(state, input, 3);
+  int64_t oheight   = THCTensor_(size)(state, grid, 1);
+  int64_t owidth    = THCTensor_(size)(state, grid, 2);
 
   THCUNN_check_dim_size(state, grid, 4, 0, nbatch);
   THCUNN_check_dim_size(state, grid, 4, 3, 2);
@@ -38,12 +38,12 @@ TH_API void THNN_(SpatialGridSamplerBilinear_updateOutput)(
 
   THCUNN_assertSameGPU(state, 3, input, grid, output);
   THNN_(SpatialGridSamplerBilinear_shapeCheck)(state, input, grid, NULL);
-  int N = THCTensor_(size)(state, input, 0);
-  int C = THCTensor_(size)(state, input, 1);
-  int IH = THCTensor_(size)(state, input, 2);
-  int IW = THCTensor_(size)(state, input, 3);
-  int H = THCTensor_(size)(state,grid, 1);
-  int W = THCTensor_(size)(state, grid, 2);
+  int64_t N = THCTensor_(size)(state, input, 0);
+  int64_t C = THCTensor_(size)(state, input, 1);
+  int64_t IH = THCTensor_(size)(state, input, 2);
+  int64_t IW = THCTensor_(size)(state, input, 3);
+  int64_t H = THCTensor_(size)(state,grid, 1);
+  int64_t W = THCTensor_(size)(state, grid, 2);
 
   // resize output to the same shape as input
   THCTensor_(resize4d)(state, output, N, C, H, W);
@@ -52,7 +52,7 @@ TH_API void THNN_(SpatialGridSamplerBilinear_updateOutput)(
   THCDeviceTensor<real, 4> devGrid = toDeviceTensor<real, 4>(state, grid);
   THCDeviceTensor<real, 4> devOutput = toDeviceTensor<real, 4>(state, output);
 
-  int count = N*H*W*2;
+  int64_t count = N*H*W*2;
   SpatialGridSamplerBilinear_updateOutput_kernel
     <<<GET_BLOCKS(count), CUDA_NUM_THREADS, 0, THCState_getCurrentStream(state)>>>(
       count, devInput, devGrid, devOutput);
@@ -66,12 +66,12 @@ TH_API void THNN_(SpatialGridSamplerBilinear_updateGradInput)(
 
   THCUNN_assertSameGPU(state, 5, input, gradInput, grid, gradGrid, gradOutput);
   THNN_(SpatialGridSamplerBilinear_shapeCheck)(state, input, grid, gradOutput);
-  int N = THCTensor_(size)(state, input, 0);
-  int C = THCTensor_(size)(state, input, 1);
-  int IH = THCTensor_(size)(state, input, 2);
-  int IW = THCTensor_(size)(state, input, 3);
-  int H = THCTensor_(size)(state, grid, 1);
-  int W = THCTensor_(size)(state, grid, 2);
+  int64_t N = THCTensor_(size)(state, input, 0);
+  int64_t C = THCTensor_(size)(state, input, 1);
+  int64_t IH = THCTensor_(size)(state, input, 2);
+  int64_t IW = THCTensor_(size)(state, input, 3);
+  int64_t H = THCTensor_(size)(state, grid, 1);
+  int64_t W = THCTensor_(size)(state, grid, 2);
 
   THCTensor_(resize4d)(state, gradInput, N, C, IH, IW);
   THCTensor_(resize4d)(state, gradGrid, N, H, W, 2);
@@ -84,7 +84,7 @@ TH_API void THNN_(SpatialGridSamplerBilinear_updateGradInput)(
   THCDeviceTensor<real, 4> devGradGrid = toDeviceTensor<real, 4>(state, gradGrid);
   THCDeviceTensor<real, 4> devGradOutput = toDeviceTensor<real, 4>(state, gradOutput);
 
-  int count = N*H*W;
+  int64_t count = N*H*W;
   SpatialGridSamplerBilinear_updateGradInput_kernel
     <<<GET_BLOCKS(count), CUDA_NUM_THREADS, 0, THCState_getCurrentStream(state)>>>(
       count, devInput, devGradInput, devGrid, devGradGrid, devGradOutput);
